Перенес логику отрисовки точки маршрута в point-controller
